Go code gen checks into pkg/generate/code
Begins the process of pulling the functionality that generates Go code
(as opposed to describing the structure of APIs and objects in an API)
into a separate `pkg/generate/code` package.

This first part of the process pulls the methods on `pkg/model:CRD` that
output Go code representing checks/assertions into a separate
`pkg/generate/code/check.go` file, along with unit tests for these
functions.

The following methods were moved:

* `CRD.GoCodeRequiredFieldsMissingFromShape` ->
  `code.CheckRequiredFieldsMissingFromShape`
* CRD.GoCodeSetExceptionMessagPrefixCheck` ->
  `code.CheckExceptionMessagePrefix`