Move SDK shape setters into pkg/generate/code
Moves all of the `pkg/model:CRD` `GoCodeSetInputXXXX` methods into
`pkg/generate/code`, removing all the "GoCode" prefixes from the method
names and replacing "SetInput" with "SetSDK" to make it more clear what
the code is doing that is produced from these functions.

The next patch changes the "SetOutput" methods to be "SetResource"
functions in the `pkg/generate/code` package to make is clear that the
code those functions produce sets fields on a CR, not an SDK shape.

Moves all unit tests of the SetInput code generation functions into a
separate file.