fix release copy file generation
The pkg/generate/ack/release.go file was calling templateset.New() with
the wrong set of copyPaths (it was passing the api copy paths not the
release copy paths). The templateset.Execute() call was also not
properly constructing a full path for copying by prepending the template
base path to the copied file path.