remove build-controllers matrix from unit test
The build-unit-test GH action job was running

```
make build-controller SERVICE=$SERVICE
```

for each service in a matrix of services. The
aws-controllers-k8s/code-generator repository will be housing the
`scripts/build-controller.sh` script from now on, and each separate
service controller repository will be getting a separate GH action job
that is triggered when PRs are pushed for either the service controller
repo or the code-generator repo. Therefore, we're removing this matrix
job from the top-level repo.

Issue #666