fix up build scripts to deal with deleted go.mod
After removing the top-level go.mod/sum files, the
`scripts/build-controller-image.sh` and
`scripts/build-controller-release.sh` scripts were broken.

In the case of `build-controller-image.sh`, we needed to always set the
Docker build context to the top-level
`$GOPATH/src/github.com/aws-controllers-k8s` directory and update the
Dockerfile to always copy the go.mod and go.sum files from the service
controller's repository instead of the old go.mod/sum files from the
original monorepo.

In the case of `build-controller-release.sh`, we needed to pass a
`--aws-sdk-go-version` CLI argument to the `ack-generate release`
command line call to force `ack-generate` to not look for a go.mod file
in order to determine the aws-sdk-go version to use in constructing the
SDK helper for the code generator. In the case of the `ack-generate
release` command, we don't actually use the aws-sdk-go contents like we
do for `ack-generate apis` and `ack-generate controller`.