Update links to building controller doc
This change fixes the links to the building controller doc in the
Overview and Setup pages of the dev docs, and adds a nav entry for
that page in the mkdocs config. Previously, these links weren't
resolving as expected on the aws-controllers-k8s.github.io site.

This change has been validated with a local mkdocs preview.