releases, project stages and maintenance phases
Our existing single "release status" was too coarse and was leading to
confusion -- both from customers as well as internal service teams --
about supportability and maintenance policy for controllers in the
different release statuses.

This patch updates our documentation to clarify what we mean by a
"release" and a "version". It also describes two separate concepts:
the **project stage** and the **maintenance phase**.

The controller's "project stage" describes how far along the controller
is towards being released:

`PROPOSED` -> `PLANNED` -> `IN PROGRESS` -> `RELEASED`

There's nothing about our support stance or what percentage of an API's
resources need to be implemented or what test coverage is the bar to
reach. Instead, the `RELEASED` project stage is reached when a SemVer
Git tag is applied to the source code repository and Docker images and
Helm Charts are automatically built and published to our ECR Public
repo.

This patch also significantly modifies our definition of what a
"release" entails as well as clarifies that we use SemVer to signal
breaking API changes.

We also clarify that there may be one and only one "stable"
artifact tag applied only to the Helm Chart for a major version series
of a controller. This "stable artifact tag" is applied by the maintainer
team after determining that a particular Docker binary image at a
specific SemVer version and configuration value combination have passed
long-running tests and the team is confident in the controller being
used in production.

Finally, there is the concept of a "maintenance phase", which describes
our support stance and the maintaining team's confidence that the
controller in a major version series is ready for production use and has
been tested in long-running operational tests.

There are four maintenance phases. A maintenance phase describes a
**major SemVer version series** of a controller in the `RELEASED` build
stage:

`PREVIEW` describes a controller major version series that is not yet
recommended for production use by the maintainer team.

`GENERALLY AVAILABLE` describes a controller major version series that
is recommended for production use by the maintainer team.

`DEPRECATED` describes a controller major version series that will no
longer be supported in the future.

`NOT SUPPORTED` describes a controller major version series that is no
longer maintained. A controller major version series can go from
`PREVIEW` directly to `NOT SUPPORTED`. However, a controller major
version series in `GENERALLY AVAILABLE` maintenance phase must go into
the `DEPRECATED` maintenance phase for one year before being moved to
`NOT SUPPORTED` maintenance phase.