add simple create-delete non-public Rabbit test
I learned quite a bit about the AmazonMQ API while putting together this
(very) simple test that creates a non-public-accessible RabbitMQ broker
and then deletes that broker after verifying it is queryable via the
AmazonMQ APIs.

The first thing I learned was that on average it takes more than 5
minutes to launch a single-instance RabbitMQ instance.

The second thing I learned was that if you create a new subnet to house
a public-accessible broker in and specify that subnet without
associating it with an internet gateway, AmazonMQ will complain that the
subnet doesn't exist even though it does exist but isn't publicly
accessible.

The third thing I learned was that you cannot delete a broker if it is
not in a RUNNING state. A broker gets to RUNNING from
CREATING_IN_PROGRESS state after 5+ minutes.

The PR I pushed to the code-generator for generic hooks will be used to
create custom callbacks that check the Status.BrokerState field on both
Update and Delete calls to ensure we don't try to update something that
is still being created or delete something that isn't yet running.

I will add additional tests for publicly-accessible broker creation as
well as update code paths for both non and publicly-accessible brokers
once I merge the generic hooks PR in the code-generator and regen the MQ
controller.