Parameterize controller build image (#820)
**Issue #, if available:**
Related to #763, but does not close this issue.

**Description of changes:**

This change adds a new `base_image` argument to all Dockerfiles, and sets the default value of this parameter to the original value of `amazonlinux:2`. This should cause no change for existing build workflows.

According to documentation, this new argument must exist prior to the very first `FROM` in order to be used in a `FROM` action call. https://docs.docker.com/engine/reference/builder/#understand-how-arg-and-from-interact

This enables developers to build a container from this Dockerfile with an alternate target base image for the final stage.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license. 👍 

Signed-off-by: Jose R. Gonzalez <josegonzalez89@gmail.com>