Fix GoDoc link in docs (#839)
Description of changes: Fix the casing for the GoDoc link, so it points to the correct API on the page.

Also updated the README to work with the structure assumed by the configuration file.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.