update docs for separate ECR repos and releases
Updates our documentation in order to have our new ECR Public
repositories exposed in appropriate links.

Removes the "Work in Progress" user documentation navigation and
adds our install, IRSA and Authorization user docs to the navigation
sidebar.

Updates the Services matrix with the most up-to-date links and statuses.

Finally, replaces the wording in the documentation index/homepage about
"Developer Preview" and "Not to use in production" with explanation
about our release process and maintenance phases and links to the
install docs.