Fix documentation website missing styles (#934)
Description of changes:
Includes a number of fixes to the documentation website, mostly regarding missing styles. These styles were purged from the end CSS by PostCSS, because they are not explicitly being used in the layouts (but are used when generating the markdown as HTML).
- Allowlist reference tables, hint boxes and code syntax highlighting in PostCSS
- Move dark mode logic to header script for faster loading (still stutters)
- Fix dark mode default dark/light theme logic

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.