Add subtle background pattern to home page (#944)
Description of changes:
Adds a subtle background pattern to the top of the home page. Fixes the container size of the html and body elements, to always be at least 100%. Sticky footer to bottom of the page.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.