Cross Region Resource Management Docs Page (#962)
Hi all! I'm continuing to work on the ACK docs and was requested to add a general Cross-Region Resource Management page because we know specific SageMaker customers are looking for this feature. The docs here are mostly from a SageMaker perspective, so I would appreciate any additional context/clarification on how to best generalize these instructions. Thank you! 

Description of changes: Added cross-region-resource-management.md to User Docs.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.