Temporarily revert Helm installation steps in SageMaker Example (#1018)
Issue #, if available:
#988 
Description of changes:
Current documentation may be confusing to customers as `SageMaker` repo hasn't released with `3.7 `helm charts yet. This upgrade is coming soon however in the meantime we want the docs to reflect the current installation process of `Sagemaker` Controller

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.