Application Auto Scaling (+SM) ACK Tutorial (#1032)
Another tutorial! This time with Application Auto Scaling and SageMaker together.

Description of changes: Tutorial that is an optional continuation of the SageMaker training tutorial or, if you already have an endpoint to scale, just an Application Auto Scaling tutorial. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.