Simplify examples in the RDS tutorial
This reduces the number of parameters required to deploy a RDS
database when following the RDS tutorial. It also uses the engine
major version number, reducing the need to continuously update the
example when a new minor release of the engine is available.

Additionally, the database username is added to the Secret for
convenience, as credentials are typically loaded in apps in this
manner.