Update docs fonts and sidebars (#1229)
Description of changes:
Something has always seemed off about the font anti-aliasing and sidebars with the website. This pull request fixes that by including all of the font weights supported by the "Open Sans" font. This dramatically improves the look of the website, by reducing the amount of anti-aliasing the browser has to do. 

This pull request also updates the styling of the sidebars to match that of the Bootstrap 5 documentation. This just improves the interaction feeling of these elements.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.