Generate runtime API reference documentation (#1226)
Description of changes:
Updates the API reference generator script to support the APIs defined in the `runtime` repository. Also updates the HTML of the API reference page to place the APIs in their own section at the top under `Common`.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.