Updating the path for the Helm Chart (#1307)
Description of changes:

When installing the ACK Resource Helm Chart there is an environment variable set as CHART_REF=$SERVICE-chart. This creates the folder under /tmp/chart as $SERVICE-CHART. The path in the commands for deleting the CRDs doesn't exist because of that. 

I've updated the page with the right path.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.