Updating the ACK Namespace environment variable (#1310)
When installing the ACK Resource Helm Chart there is an environment variable set as ACK_SYSTEM_NAMESPACE=ack-system. The given variable for deletion is different than what was set before.

I've updated the environment variable with the right one.

Issue #, if available:

Description of changes:

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.