Update Application Autoscaling tutorial instance type (#1633)
Issue #, if available:

Sagemaker endpoint no longer supports "t2" type instances for autoscaling, an instance of that type is currently being used in [this](https://aws-controllers-k8s.github.io/community/docs/tutorials/autoscaling-example/#deploy-a-sagemaker-endpoint) tutorial. This PR replaces it with a ml.m5.large.

Description of changes:

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.