Sort table of contents by short name (#1638)
Issue #, if available: https://github.com/aws-controllers-k8s/community/issues/1637

Description of changes:
Currently, when the scripts that generate the API reference execute, the docs are generated using the alphabetical order of the controller repositories. This places the `prometheusservice-controller` after the `organizations-controller` in the API reference table of contents. However, when rendering the ToC, they are rendered using their short name. This would require that `AMP` is after `ACM` and before `APIGWv2` in the list. This pull request sorts the list of services by their metadata "short name" before executing their templates and adding the data to the YAML.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.