Improve documentation website UX (#1751)
As we added more tutorials and new services to ACK and our documentation
website, it got harder to navigate throught the docs menu. This update
improves the UX by tweaking the `docs-menu.html` and
`reference-menu-item.html` files to make it easier to use the pages
sections.

This patch bring mainly two changes:
- On the `Documentation` page, all the sections (except the active one)
  are collapsed by default. No more getting lost in a sea of sections!
- On the `API References` page, the expand icon now correctly shows the
  status of the section/subsection (expanded/unexpanded), when selecting
  a new CRD.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.