Add drift remediation configuration document (#1811)
Issue #, if available:
Closes https://github.com/aws-controllers-k8s/community/issues/1810

Description of changes:
Adds a new document to the "Getting Started" tab of the documentation which outlines the definition and configuration options for ACK drift remediation. Also adds a link to it from the "How it Works" page.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.