Update irsa.md (#2217)
Point to newer docs for using AWS IAM Roles Anywhere.

Issue #, if available:
N/A 
Description of changes:
For non EKS clusters Digizens have been pointed to older documentation that needed to be updated. 

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.